Fikse at kanalform ikke står i capslock + rename aktivitetTypes
