+++ src/utils/aktivitetTypes.ts
