+++ src/view/AktivitetProvider.tsx
@@ -1,7 +1,7 @@
 import useFetch, { FetchResult, Status, hasData, hasError } from '@nutgaard/use-fetch';
 import React, { useContext } from 'react';
 
-import { Aktivitet, ArenaAktivitet } from '../utils/AktivitetTypes';
+import { Aktivitet, ArenaAktivitet } from '../utils/aktivitetTypes';
 import { REQUEST_CONFIG, fnrQuery } from '../utils/Fetch';
 import { StringOrNull } from '../utils/Typer';
