+++ src/view/aktivitet/MoteAktivitet.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Aktivitet, KanalTypes } from '../../utils/AktivitetTypes';
+import { Aktivitet, KanalTypes } from '../../utils/aktivitetTypes';
 import { formaterDate, getKlokkeslett, getVarighet } from '../../utils/Date';
 import InformasjonElement from './InformasjonElement';
 
@@ -10,12 +10,11 @@ interface PropTypes {
 
 export default function MoteAktivitet(props: PropTypes) {
     const { fraDato, tilDato, kanal, adresse, beskrivelse, forberedelser } = props.aktivitet;
-
     return (
         <>
             <InformasjonElement merkelapptekst="Dato" verdi={formaterDate(fraDato)} />
             <InformasjonElement merkelapptekst="Klokkeslett" verdi={getKlokkeslett(fraDato)} />
-            <InformasjonElement merkelapptekst="Kanal" verdi={getText(kanal)} />
+            <InformasjonElement merkelapptekst="Kanal" verdi={getKanalTekst(kanal)} />
             <InformasjonElement merkelapptekst="Varighet" verdi={getVarighet(fraDato, tilDato)} />
             <InformasjonElement merkelapptekst="Adresse" verdi={adresse} />
             <InformasjonElement merkelapptekst="Bakgrunn" verdi={beskrivelse} />
@@ -24,7 +23,7 @@ export default function MoteAktivitet(props: PropTypes) {
     );
 }
 
-function getText(kanal: KanalTypes) {
+export const getKanalTekst = (kanal: KanalTypes) => {
     switch (kanal) {
         case KanalTypes.OPPMOTE:
             return 'Oppmøte';
@@ -32,6 +31,7 @@ function getText(kanal: KanalTypes) {
             return 'Telefonmøte';
         case KanalTypes.INTERNETT:
             return 'Nettmøte';
+        default:
+            return '';
     }
-    return '';
-}
+};
